+++ vector/src/main/scala/geotrellis/vector/GeomFactory.scala
@@ -25,9 +25,9 @@ import com.vividsolutions.jts.precision.GeometryPrecisionReducer
 
 import scala.util.Try
 
-private[vector] object GeomFactory extends LazyLogging {
+object GeomFactory extends LazyLogging {
 
-  val precisionType = {
+  private[vector] val precisionType = {
     val setting = Try(ConfigFactory.load().getString("geotrellis.jts.precision.type").toLowerCase)
     if (setting.isSuccess)
       setting.get
@@ -37,14 +37,16 @@ private[vector] object GeomFactory extends LazyLogging {
     }
   }
 
-  val precisionModel = precisionType match {
+  private[vector] lazy val fixedPrecisionScale =
+    Try(ConfigFactory.load().getDouble("geotrellis.jts.precision.scale"))
+
+  private[vector] val precisionModel = precisionType match {
     case "floating" => new PrecisionModel()
     case "floating_single" => new PrecisionModel(PrecisionModel.FLOATING_SINGLE)
     case "fixed" =>
-      val scaleFromConfig = Try(ConfigFactory.load().getDouble("geotrellis.jts.precision.scale"))
       val scale =
-        if (scaleFromConfig.isSuccess)
-          scaleFromConfig.get
+        if (fixedPrecisionScale.isSuccess)
+          fixedPrecisionScale.get
         else {
           logger.warn("No value specified in application.conf for geotrellis.jts.precision.scale; using default")
           1e12
@@ -53,10 +55,15 @@ private[vector] object GeomFactory extends LazyLogging {
     case s => throw new IllegalArgumentException(s"""Unrecognized JTS precision model, ${precisionType}; expected "floating", "floating_single", or "fixed" """)
   }
 
+  /** The JTS GeometryFactory used to create new geometries.
+   *
+   *  Use this factory object to create geometries that avoid automatic snapping
+   *  of geometry coordinates to the current precision model.
+   */
   val factory = new geom.GeometryFactory(precisionModel)
 
   // 12 digits is maximum to avoid [[TopologyException]], see https://web.archive.org/web/20160226031453/http://tsusiatsoftware.net/jts/jts-faq/jts-faq.html#D9
-  lazy val simplifier = {
+  private[vector] lazy val simplifier = {
     val simplificationPrecision = Try(ConfigFactory.load().getDouble("geotrellis.jts.simplification.scale"))
     val scale =
       if (simplificationPrecision.isSuccess)